backend/requirements.txt
# Core Flask
Flask==2.0.3
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Werkzeug==2.0.3
SQLAlchemy==1.4.46

# Authentication
PyJWT==2.4.0
email-validator==1.1.3
python-dotenv==0.20.0

# Image Processing
opencv-python==4.5.5.64
Pillow==8.4.0
numpy==1.19.5
scikit-image==0.17.2

# Machine Learning
torch==1.10.2+cu113 -f https://download.pytorch.org/whl/cu113/torch_stable.html
torchvision==0.11.3+cu113 -f https://download.pytorch.org/whl/cu113/torch_stable.html
ultralytics>=8.0.0

# Data handling
pandas==1.3.5
requests==2.31.0

# File handling
python-multipart==0.0.6